requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
